pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "fastapi>=0.115.11",
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "flask-sqlalchemy>=3.1.1",
    "langchain>=0.3.20",
    "langchain-community>=0.3.19",
    "langchain-openai>=0.3.8",
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.1",
    "pypdf>=5.3.1",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "pytz>=2025.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.39",
    "sqlalchemy-utils>=0.41.2",
    "uvicorn>=0.34.0",
]
